+++ examples/documentation/quick/basic/source/validate-utf32.cpp
@@ -1,7 +1,7 @@
 // =============================================================================
 //
 // ztd.text
-// Copyright © 2022 JeanHeyd "ThePhD" Meneide and Shepherd's Oasis, LLC
+// Copyright © 2022-2023 JeanHeyd "ThePhD" Meneide and Shepherd's Oasis, LLC
 // Contact: opensource@soasis.org
 //
 // Commercial License Usage
@@ -18,7 +18,7 @@
 // file except in compliance with the License. You may obtain a copy of the
 // License at
 //
-//		http://www.apache.org/licenses/LICENSE-2.0
+// https://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -61,7 +61,8 @@ int main(int, char*[]) {
 	std::u32string_view unused_input(
 	     validate_result.input.data(), validate_result.input.size());
 	for (const auto& u32_codepoint : unused_input) {
-		std::cout << "0x" << std::hex << u32_codepoint;
+		std::cout << "0x" << std::hex
+		          << static_cast<uint_least32_t>(u32_codepoint);
 		if (&u32_codepoint != &unused_input.back()) {
 			std::cout << " ";
 		}
